Removal of German Language
**Removal of German Language**

In Germany, internet filtering and violations of users' internet freedoms do not exist in a manner similar to Iran, China, or Russia. The goal of this program has never been to include all languages. Especially in countries where government surveillance and filtering are used to combat cybercrime, rather than to restrict public access to the internet, English is sufficient as the standard language for those countries.